+++ modules/output/__init__.py
@@ -1,0 +1,3 @@
+class OutputModule(object):
+    def output(self, stats):
+        pass
