--- modules/output/output.py
@@ -1,3 +1,0 @@
-class OutputModule(object):
-    def output(self, stats):
-        pass
